+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Current changes without release yet
 
+- Repaired error #91 adding a force-overwrite option to the command line
+
 ## [0.1.6] - 2024-08-09
 
 - Added more features to SHACL validation: #94